+++ src/main.ts
@@ -1,4 +1,6 @@
 import { bootstrap } from "@app/app"; // import the function from app.ts
+import env from "@config/env"; 		  // import the env object from env.ts
+
 import express from "express";
 
 const init = async () => {
@@ -9,8 +11,8 @@ const init = async () => {
 	bootstrap(app);
 
 	// Start server and listen requests on port 3005
-	app.listen(3000, () => {
-		console.log('Server is running on port 3005');
+	app.listen(env.PORT, () => {
+		console.log(`Server is running on port ${env.PORT}`);
 	});
 }