+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # CHANGELOG
 
 ## Initial setup
-> Current
 
 - Created project
 - Configured typescript
 - Configured Prettier formatter
 - Configured eslint
 - Added bootstrap code with test endpoint `/`
+
+## Dotenv config
+> Current
+
+- Now the project reads `.env` files
+- Added a `.env.example`